+++ src/asora/raytracing.cuh
@@ -36,7 +36,7 @@ void do_all_sources_gpu(
 
 // Raytracing kernel, called by do_all_sources
 __global__ void evolve0D_gpu(
-    const int Rmax_LLS,
+    const double Rmax_LLS,
     const int q,
     const int ns_start,
     const int num_src_par,
